fixed and added functionality to deleteById method for manager entity

+++ Employee-Management-System/src/main/java/com/capg/controller/ManagerController.java
@@ -1,8 +1,8 @@
 package com.capg.controller;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,13 +15,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.capg.entity.Manager;
 import com.capg.entity.Employee;
+import com.capg.entity.Manager;
 import com.capg.exception.EnterValidDetailsException;
 import com.capg.exception.ManagerEmptyException;
 import com.capg.exception.ManagerNotFoundException;
 import com.capg.exception.ProjectNotFoundException;
 import com.capg.service.ManagementService;
+import com.google.common.base.Predicate;
 
 @RestController
 @RequestMapping("/managers")
@@ -76,17 +77,17 @@ public class ManagerController {
 	
 	@DeleteMapping("/delete-by-id/{managerId}")
     public void deleteByid(@PathVariable int managerId) {
-		
-		
-		List<Employee> list=managementService.getAllEmployees();
-		for(Employee employee: list) {
-			
-			if(employee.getManager().getId()==managerId) {		
-			throw new EnterValidDetailsException("You Cant Delete This Manager Id As this Manager Is Assigned to Employees ");
-			}
-			
-		}
-		
+//		
+//		
+//		List<Employee> list=managementService.getAllEmployees();
+//		for(Employee employee: list) {
+//			
+//			if(employee.getManager().getId()==managerId) {		
+//			throw new EnterValidDetailsException("You Cant Delete This Manager Id As this Manager Is Assigned to Employees ");
+//			}
+//			
+//		}
+//		
 		if(managerId<0) {		
 			throw new EnterValidDetailsException("Please Enter Valid Manager Id");
 		}
@@ -95,10 +96,17 @@ public class ManagerController {
     	if (!manager.isPresent()) {
 			throw new ManagerNotFoundException("Does not found manager with " + managerId);
 		}
-//    	manager.get().getEmployees().stream().forEach((p) -> p.setManager(null));
-//    	manager.get().getProject().setManager(null);
-//    	managementService.deleteByManagerId(managerId);
-    	  managementService.deleteByManagerId(managerId);
+    	Predicate<Manager> isProjectPresent = (m) -> m.getProject() != null;
+    	Predicate<Manager> areEmployeesPresent = (m) -> !(m.getEmployees()==null || m.getEmployees().isEmpty());
+    	if(isProjectPresent.apply(manager.get()) & areEmployeesPresent.apply(manager.get())) {
+    		throw new EnterValidDetailsException("You cant delete this manager with id " + manager.get().getId() +  " as this manager is assigned to both employee(s) with id(s) " + manager.get().getEmployees().stream().map((e) -> String.valueOf(e.getId())).collect(Collectors.joining(",")) + " and project with id " + manager.get().getProject().getId());
+    	} else if(isProjectPresent.apply(manager.get())) {
+    		throw new EnterValidDetailsException("You cant delete this manager with id " + manager.get().getId() +  " as this manager is assigned to project with id " + manager.get().getProject().getId());
+    	}else if(areEmployeesPresent.apply(manager.get())) {
+    		throw new EnterValidDetailsException("You cant delete this manager with id " + manager.get().getId() +  " as this manager is assigned to employee(s) with id(s) - " + manager.get().getEmployees().stream().map((e) -> String.valueOf(e.getId())).collect(Collectors.joining(",")));
+    	} else {
+    		managementService.deleteByManagerId(managerId);
+    	}
     }
  }
 }
